% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_logger.R
\name{fn_logger}
\alias{fn_logger}
\title{Logger}
\usage{
fn_logger(msg, level = "info", file = log_file)
}
\arguments{
\item{msg}{message text}

\item{level}{classification of the message}

\item{file}{file to write to}
}
\value{
None
}
\description{
Creates a logger function
}
\examples{
fn_logger("Hello world")

}

