% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_lr_backwards_elimination.R
\name{fn_lr_backwards_elimination}
\alias{fn_lr_backwards_elimination}
\title{Automated Backwards Elimination for Logistic Regression}
\usage{
fn_lr_backwards_elimination(lr_model)
}
\arguments{
\item{lr_model}{logistic regression object created by glm}
}
\value{
updated logistic regression object
}
\description{
Performs backwards elimination on a logistic regression model
}
\examples{


}

