% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot_cv_glmnet.R
\name{fn_plot_cv_glmnet}
\alias{fn_plot_cv_glmnet}
\title{Plot of cv.glmnet objects}
\usage{
fn_plot_cv_glmnet(cv_glmnet, main, users_v = c())
}
\arguments{
\item{cv_glmnet}{cv.glmnet object}

\item{main}{title of the plot}

\item{users_v}{character vector of users if fonts installed}
}
\value{
None
}
\description{
Creates a styled plot of cv.glmnet objects
}
\examples{
fn_plot_cv_glmnet(cv_glmnet, main = "LASSO")

}

