% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_regsubsets_plots.R
\name{fn_regsubsets_plots}
\alias{fn_regsubsets_plots}
\title{Regsubsets Plots}
\usage{
fn_regsubsets_plots(regsubsets_obj, elbow = NULL, users_v = c())
}
\arguments{
\item{regsubsets_obj}{object created by regsubsets object}

\item{elbow}{xintercept (# of predictors) to show an elbow}

\item{users_v}{character vector of users}
}
\value{
ggplot object
}
\description{
Creates plots of Cp, BIC, and R^2 from a regsubsets object
}
\examples{
fn_regsubsets_plots(regsubsets_obj, users_v = "Jordan")

}

