% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal538.R
\docType{data}
\name{pal538}
\alias{pal538}
\title{Fivethirtyeight theme colors}
\format{An object of class \code{character} of length 6.}
\usage{
pal538
}
\value{
None
}
\description{
Creates a named character vector of Fivethirtyeight's colors
}
\examples{
pal538()

}
\keyword{datasets}

