% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jrf.R
\name{theme_jrf}
\alias{theme_jrf}
\title{Fivethirtyeight theme for ggplot2}
\usage{
theme_jrf(base_size = 8, users_v = c())
}
\arguments{
\item{base_size}{base font size, defaults to 8}

\item{users_v}{character vector of users}
}
\value{
None
}
\description{
Creates a ggplot2 theme function
}
\examples{
ggplot(mtcars, aes(weight, mpg)) + geom_point() + theme_jrf()

}

